Update perl RPM with HV logic and use it
Plug declare_shared_memory_usage calls
to declare on demand any reserved memory.

At this stage PL_static_shared_memory_table
is one HV that contains all the size for the memory
address of all our entries.

The HV is here as a proof of concept mainly
and is going to be improved to use a static
declaration later.