[idea] Proof of concept to use cowpv for shared_hek
Works on a basic q{hello world} program

This commit is doom by definition of shared_heks.
It's going to explode the second we use the flags
from the shared_hek.

- from the PV you need to point to the beginning of the struct
( view SvSHARED_HEK_FROM_PV )
- from the string you need to access to the flag stored just
at the end

Note that the COWPV string cannot be stored outside of the
struct as done here.

The COWPV definition would need to point to shared_hek one
and the flag woyld need to be moved in a different location...

This is tricky as ideally you would like to know all the COWPVs
before saving the shared_heks, but then you would also need
to remove them from the master-cow-pv....