Fix SEGV when clearing UNOP_AUX with PL_*
Case BC-3093:

Avoid to free CORE GVs when clearing UNOP_AUX.

In order to use %ENV and %ARGV a compiled program
will directly use PL_envgv or PL_argvgv, but do not hold
a refcnt on it. So when destroying the UNOP_AUX this could
result in clearing the GV itself..

We should increase the refcnt when using these variables.

The issue was limited to PL_envgv and PL_argvgv
listed as special CORE_SYMS in GV.pm

Add unit tests to show the SEGV issue when
clearing UNOP_AUX using %ENV or %ARGV.